day 22 part 2
at last.
Code is awful, lot of dead useless code,
but I don't care and keep it for history purpose.

I found it mostly thanks to a detailed algorithmic solution on the web,
so I wrote a mini-writeup in the notes.txt to be certain
I fully understood what I coded.

I hope the end, and 2020 edition will be less harsh from now on.